+++ testUtils/mock.js
@@ -1,0 +1,37 @@
+import { MissionUtils } from '@woowacourse/mission-utils';
+import { EOL as LINE_SEPARATOR } from 'os';
+
+const mockQuestions = (inputs) => {
+  MissionUtils.Console.readLineAsync = jest.fn();
+
+  MissionUtils.Console.readLineAsync.mockImplementation(() => {
+    const input = inputs.shift();
+
+    return Promise.resolve(input);
+  });
+};
+
+const getLogSpy = () => {
+  const logSpy = jest.spyOn(MissionUtils.Console, 'print');
+  logSpy.mockClear();
+
+  return logSpy;
+};
+
+const expectLogContains = (received, expectedLogs) => {
+  expectedLogs.forEach((log) => {
+    expect(received).toContain(log);
+  });
+};
+
+const getOutput = (logSpy) => {
+  return [...logSpy.mock.calls].join(LINE_SEPARATOR);
+};
+
+export {
+  expectLogContains,
+  getLogSpy,
+  getOutput,
+  mockQuestions,
+  LINE_SEPARATOR,
+};
