+++ testUtils/orderList.js
@@ -1,0 +1,5 @@
+import { OrderedMenu } from '../src/models';
+
+const getOrderList = (array) => array.map((v) => new OrderedMenu(v).getData());
+
+export { getOrderList };
